Fixed major bug with specifying user-defined surface reflectances.
This involved a number of fairly significant changes:

* All wavelength constants have changed from a simple negative number
to a tuple which is (6S ID, start wavelength, end wavelength)

* All functions that deal with these wavelength constants (the Wavelength function and the all_wavelengths class)
have been modified to deal with this change of constants.

* The ground reflectance Homog Lambertian and Hetero Lambertian functions now deal
with the user-defined reflectances properly, and add in the wavelength range as an extra line when needed. This
is done by returning a tuple from the Wavelength function, which is (input file string, start wavelength, stop wavelength)
and this is then inserted into the string REPLACEHERE for the GroundReflectance results.

* The centre wavelengths in the all_wavelengths functions (eg. run_landsat_etm) are now calculated from
the max and min wavelengths.