Updated ground reflectance code to allow spectral library use
The HomogenousLambertian function now allows a ndarray to be given,
which should be 2 columns (wavelength, reflectance), and then whenever
the model is run this is resampled (and/or interpolated) to give
the reflectance for the wavelength or wavelength range being used for that run.

Has been checked with vegetation spectra and seems to be doing sensible things.
No tests for any files with missing data yet though.