Fixed another bug in the AERONET import routine.
It used to crash when the nearest date/time had any NaN values for
important data (eg. Refractive Indices). It now gets the nearest time,
checks for NaNs, and if there are any, gradually works its way down
the 'nearest indices to the given time' list until it gets a row without
NaNs.