Added exception handling for 6S output parsing
Now whenever it is converting something from anywhere in the output file to put it in the outputs structure, it wraps it in a try/catch block to catch things that won't convert - eg. -1#IND000 which seems to be the way that NaN's are represented in the
output.

Deals with a bug where Py6S crashes in the SWIR under certain conditions, as reported by Martin Black