Update outputs.py
The invalid values in the output text file are represented by "********" and * should not removed from the output str, since that will results in wrong index of results. In the Version 2 6S, the results also have xap, xb and xc for the correction of reflectance, in which xap overwrites xa causing wrong correction parameters for radiance correction.