Fixed bug with the way that REFL_REPLACE and REFL_REPLACE_2 work for HeterogeneousLambertian
If the HeterogeneousLambertian GroundReflectance parameterisation is chosen, then two sets of reflectances need to be
written into the 6S input file. If these reflectances are arrays of wavelengths and reflectances (of the shape (n, 2))
then these need processing at a later stage once the wavelength information is known (which it isn't at the time that
the GroundReflectance class is instantiated). This is handled by putting two placeholder strings into the output string
from the GroundReflectance class: REFL_REPLACE and REFL_REPLACE_2. However, when replacing these strings (later on in
the process), we replaced REFL_REPLACE first, which overwrote REFL_REPLACE_2 because the string REFL_REPLACE is in both
of them. By simply reversing the order of replacing REFL_REPLACE and REFL_REPLACE_2 we sort this problem.

I really should add a test to allow the CI to check this is working!